feat: Add webbased VNC viewer (#306)

+++ src/install.sh
@@ -1,16 +1,6 @@
 #!/usr/bin/env bash
 set -Eeuo pipefail
 
-BASE="boot.img"
-[ -f "$STORAGE/$BASE" ] && return 0
-
-if [ -z "$BOOT" ]; then
-  error "No boot disk specified, set BOOT= to the URL of an ISO file." && exit 64
-fi
-
-BASE=$(basename "$BOOT")
-[ -f "$STORAGE/$BASE" ] && return 0
-
 # Check if running with interactive TTY or redirected to docker log
 if [ -t 1 ]; then
   PROGRESS="--progress=bar:noscroll"
@@ -33,6 +23,16 @@ if [[ "${BOOT_MODE,,}" == "windows" ]]; then
 
   fi
 fi
+
+BASE="boot.img"
+[ -f "$STORAGE/$BASE" ] && return 0
+
+if [ -z "$BOOT" ]; then
+  error "No boot disk specified, set BOOT= to the URL of an ISO file." && exit 64
+fi
+
+BASE=$(basename "$BOOT")
+[ -f "$STORAGE/$BASE" ] && return 0
 
 TMP="$STORAGE/${BASE%.*}.tmp"
 rm -f "$TMP"
